Add login_required decorators to chosen views

+++ products/views.py
@@ -1,6 +1,7 @@
 """ File specyfying what views can be found in the app. """
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Book, Genre, Review
@@ -76,6 +77,7 @@ def book_details(request, book_id):
     return render(request, 'products/book_details.html', context)
 
 
+@login_required
 def delete_review(request, review_id):
     """ View allowing users do delete existing reviews. """
     review = Review.objects.get(pk=review_id)
@@ -91,8 +93,13 @@ def delete_review(request, review_id):
     return redirect('account_login')
 
 
+@login_required
 def add_book(request):
     """ Add a book to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES)
         if form.is_valid():
@@ -114,8 +121,13 @@ def add_book(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_book(request, book_id):
     """ Edit a book in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+        
     book = get_object_or_404(Book, pk=book_id)
     if request.method == 'POST':
         form = BookForm(request.POST, request.FILES, instance=book)
@@ -140,8 +152,13 @@ def edit_book(request, book_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_book(request, book_id):
     """ Delete a book from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that!')
+        return redirect(reverse('home'))
+        
     book = get_object_or_404(Book, pk=book_id)
     book.delete()
     messages.success(request, 'Book deleted!')
